tests/integrations/requirements.txt
# Core testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Environment and configuration
python-dotenv>=1.0.0
PyYAML>=6.0

# Image processing
Pillow>=9.0.0

# HTTP requests for debugging
requests>=2.28.0

# Type hints
typing-extensions>=4.0.0

# Optional: For better test reporting
pytest-html>=3.1.0
pytest-cov>=4.0.0

# AI/ML SDK dependencies
openai>=1.30.0
anthropic>=0.25.0
litellm>=1.35.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-mistralai>=0.1.0
langgraph>=0.1.0
mistralai>=0.4.0
google-genai>=1.0.0

# Optional testing utilities
httpx>=0.25.0
pytest-timeout>=2.1.0
pytest-mock>=3.11.0

# Development dependencies (optional)
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting
mypy>=1.5.0   # Type checking 